requirements.txt
art~=6.1
aiofiles~=23.2.1
pyuseragents~=1.0.5
loguru~=0.7.2
web3~=6.17.2
curl_cffi~=0.6.2
PyYAML~=6.0.1
pydantic~=2.7.0
httpx~=0.27.0